Moves product buy button
